# onyx-procurement/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ONYX Procurement API
  description: |
    API for the ONYX Procurement System — covers procurement (suppliers,
    RFQ, quotes, POs, subcontractors), VAT (PCN836), annual tax forms,
    bank reconciliation, payroll/wage slips, and WhatsApp webhooks.
  version: 1.0.0
  contact:
    name: Techno-Kol Ops
servers:
  - url: http://localhost:3100
    description: Local development
  - url: https://api.onyx.example.com
    description: Production placeholder
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Employer:
      type: object
      additionalProperties: true
    Employee:
      type: object
      additionalProperties: true
    WageSlip:
      type: object
      additionalProperties: true
    EmployeeBalance:
      type: object
      additionalProperties: true
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: State conflict (already decided/closed/etc.)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /api/payroll/employers:
    get:
      tags: [Payroll]
      summary: List employers
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  employers:
                    type: array
                    items: { $ref: "#/components/schemas/Employer" }
    post:
      tags: [Payroll]
      summary: Create employer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Employer" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  employer: { $ref: "#/components/schemas/Employer" }
  /api/payroll/employees:
    get:
      tags: [Payroll]
      summary: List employees
      parameters:
        - in: query
          name: employer_id
          schema: { type: string }
        - in: query
          name: active
          schema: { type: string, enum: ["true"] }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  employees:
                    type: array
                    items: { $ref: "#/components/schemas/Employee" }
    post:
      tags: [Payroll]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Employee" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee: { $ref: "#/components/schemas/Employee" }
  /api/payroll/employees/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [Payroll]
      summary: Update employee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Employee" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee: { $ref: "#/components/schemas/Employee" }
  /api/payroll/wage-slips:
    get:
      tags: [Payroll]
      summary: List wage slips
      parameters:
        - in: query
          name: employer_id
          schema: { type: string }
        - in: query
          name: employee_id
          schema: { type: string }
        - in: query
          name: period_year
          schema: { type: integer }
        - in: query
          name: period_month
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 200 }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  wage_slips:
                    type: array
                    items: { $ref: "#/components/schemas/WageSlip" }
    post:
      tags: [Payroll]
      summary: Compute and persist a wage slip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employee_id, period]
              properties:
                employee_id: { type: integer }
                period:
                  type: object
                  properties:
                    year: { type: integer }
                    month: { type: integer }
                timesheet:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  wage_slip: { $ref: "#/components/schemas/WageSlip" }
        "409":
          description: Duplicate wage slip for this period
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/payroll/wage-slips/compute:
    post:
      tags: [Payroll]
      summary: Preview a wage slip (no persistence)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employee_id]
              properties:
                employee_id: { type: integer }
                period:
                  type: object
                  additionalProperties: true
                timesheet:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Preview computed
          content:
            application/json:
              schema:
                type: object
                properties:
                  wage_slip: { $ref: "#/components/schemas/WageSlip" }
                  preview: { type: boolean }
  /api/payroll/wage-slips/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [Payroll]
      summary: Get one wage slip
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  wage_slip: { $ref: "#/components/schemas/WageSlip" }
  /api/payroll/wage-slips/{id}/approve:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      tags: [Payroll]
      summary: Approve a wage slip
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "409": { $ref: "#/components/responses/Conflict" }
  /api/payroll/wage-slips/{id}/issue:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      tags: [Payroll]
      summary: Generate PDF and mark issued
      responses:
        "200":
          description: PDF generated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/payroll/wage-slips/{id}/pdf:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [Payroll]
      summary: Download wage slip PDF
      responses:
        "200":
          description: PDF stream
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/payroll/wage-slips/{id}/void:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      tags: [Payroll]
      summary: Void a wage slip
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Voided
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/payroll/employees/{id}/balances:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [Payroll]
      summary: Latest balance snapshot
      responses:
        "200":
          description: Balances
          content:
            application/json:
              schema:
                type: object
                properties:
                  balances: { $ref: "#/components/schemas/EmployeeBalance" }
    post:
      tags: [Payroll]
      summary: Upsert balance snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeBalance" }
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  balances: { $ref: "#/components/schemas/EmployeeBalance" }
...